ROS2 sound transmission no longer employs Images, and use the sound construct instead
