+++ wrapyfi/middlewares/ros2.py
@@ -14,13 +14,13 @@ class ROS2Middleware(metaclass=SingletonOptimized):
         ROS2Middleware(**kwargs)
 
     def __init__(self, *args, **kwargs):
-        logging.info("Initialising ROS 2 middleware")
+        logging.info("Initialising ROS2 middleware")
         rclpy.init(args=[*args], **kwargs)
         atexit.register(MiddlewareCommunicator.close_all_instances)
         atexit.register(self.deinit)
 
     @staticmethod
     def deinit():
-        logging.info("Deinitialising ROS 2 middleware")
+        logging.info("Deinitialising ROS2 middleware")
         rclpy.try_shutdown()
