+++ wrapyfi/publishers/ros2.py
@@ -241,7 +241,7 @@ class ROS2AudioChunkPublisher(ROS2Publisher):
     def __init__(self, name: str, out_topic: str, should_wait: bool = True, queue_size: int = QUEUE_SIZE,
                  channels: int = 1, rate: int = 44100, chunk: int = -1, **kwargs):
         """
-        The AudioChunkPublisher using the ROS2 Image message assuming a numpy array as input
+        The AudioChunkPublisher using the ROS2 Audio message assuming a numpy array as input
 
         :param name: str: Name of the publisher
         :param out_topic: str: Name of the output topic preceded by '/' (e.g. '/topic')
@@ -251,8 +251,7 @@ class ROS2AudioChunkPublisher(ROS2Publisher):
         :param rate: int: Sampling rate. Default is 44100
         :param chunk: int: Chunk size. Default is -1 meaning that the chunk size is not fixed
         """
-        super().__init__(name, out_topic, should_wait=should_wait, queue_size=queue_size,
-                         width=chunk, height=channels, rgb=False, fp=True, jpg=False, **kwargs)
+        super().__init__(name, out_topic, should_wait=should_wait, queue_size=queue_size, **kwargs)
 
         self.channels = channels
         self.rate = rate
@@ -270,8 +269,17 @@ class ROS2AudioChunkPublisher(ROS2Publisher):
         :param repeats: int: Number of repeats to await connection. None for infinite. Default is None
         :return: bool: True if connection established, False otherwise
         """
-        # self._publisher = rospy.Publisher(self.out_topic, sensor_msgs.msg.Image, queue_size=self.queue_size)
-        self._publisher = self.create_publisher(sensor_msgs.msg.Image, self.out_topic, qos_profile=self.queue_size)
+        try:
+            from wrapyfi_ros2_interfaces.msg import ROS2AudioMessage
+        except ImportError:
+            import wrapyfi
+            logging.error("[ROS2] Could not import ROS2AudioMessage. "
+                          "Make sure the ROS2 services in wrapyfi_extensions/wrapyfi_ros2_interfaces are compiled. "
+                          "Refer to the documentation for more information: \n" +
+                          wrapyfi.__url__ + "wrapyfi_extensions/wrapyfi_ros2_interfaces/README.md")
+            sys.exit(1)
+        self._publisher = self.create_publisher(ROS2AudioMessage, self.out_topic, qos_profile=self.queue_size)
+        self._sound_msg = ROS2AudioMessage()
         established = self.await_connection(self._publisher)
         return self.check_establishment(established)
 
@@ -300,14 +308,15 @@ class ROS2AudioChunkPublisher(ROS2Publisher):
             raise ValueError("Incorrect audio shape for publisher")
         aud = np.require(aud, dtype=np.float32, requirements='C')
 
-        aud_msg = sensor_msgs.msg.Image()
+        aud_msg = self._sound_msg
         aud_msg.header.stamp = self.get_clock().now().to_msg()
-        aud_msg.height = chunk
-        aud_msg.width = channels
-        aud_msg.encoding = '32FC1'
+        aud_msg.chunk_size = chunk
+        aud_msg.channels = channels
+        aud_msg.sample_rate = rate
         aud_msg.is_bigendian = aud.dtype.byteorder == '>' or (aud.dtype.byteorder == '=' and sys.byteorder == 'big')
+        aud_msg.encoding = 'S16BE' if aud_msg.is_bigendian else 'S16LE'
         aud_msg.step = aud.strides[0]
-        aud_msg.data = aud.tobytes()
+        aud_msg.data = aud.tobytes()  # (aud * 32767.0).tobytes()
         self._publisher.publish(aud_msg)
