+++ wrapyfi_extensions/wrapyfi_ros2_interfaces/srv/ROS2AudioService.srv
@@ -1,0 +1,3 @@
+string request
+---
+wrapyfi_ros2_interfaces/ROS2AudioMessage response
